Wrap PG-specific integration with dry-c extension
