+++ lib/rom/sql/commands.rb
@@ -3,6 +3,3 @@ require 'rom/commands'
 require 'rom/sql/commands/create'
 require 'rom/sql/commands/update'
 require 'rom/sql/commands/delete'
-
-require 'rom/sql/commands/postgres'
-require 'rom/sql/commands_ext/postgres'
