+++ lib/rom/sql/extensions/active_support_notifications.rb
