+++ lib/rom/sql/extensions/postgres.rb
@@ -1,0 +1,2 @@
+require 'rom/sql/extensions/postgres/commands'
+require 'rom/sql/extensions/postgres/types'
