+++ lib/rom/sql/extensions/postgres/types.rb
@@ -8,18 +8,18 @@ module ROM
         Sequel.extension(:pg_json)
 
         Array = Dry::Types::Definition
-          .new(Sequel::Postgres::JSONArray)
-          .constructor(Sequel.method(:pg_json))
+                .new(Sequel::Postgres::JSONArray)
+                .constructor(Sequel.method(:pg_json))
 
         Hash = Dry::Types::Definition
-          .new(Sequel::Postgres::JSONHash)
-          .constructor(Sequel.method(:pg_json))
+               .new(Sequel::Postgres::JSONHash)
+               .constructor(Sequel.method(:pg_json))
 
         JSON = Array | Hash
 
         Bytea = Dry::Types::Definition
-          .new(Sequel::SQL::Blob)
-          .constructor(Sequel::SQL::Blob.method(:new))
+                .new(Sequel::SQL::Blob)
+                .constructor(Sequel::SQL::Blob.method(:new))
       end
     end
   end
