+++ lib/rom/sql/extensions/rails_log_subscriber.rb
