+++ spec/unit/types_spec.rb
@@ -1,8 +1,7 @@
 require 'rom/sql/types'
-require 'rom/sql/types/pg'
 
-RSpec.describe ROM::SQL::Types, :postgres do
-  describe ROM::SQL::Types::Serial do
+RSpec.describe 'ROM::SQL::Types', :postgres do
+  describe 'ROM::SQL::Types::Serial' do
     it 'accepts ints > 0' do
       expect(ROM::SQL::Types::Serial[1]).to be(1)
     end
@@ -12,7 +11,7 @@ RSpec.describe ROM::SQL::Types, :postgres do
     end
   end
 
-  describe ROM::SQL::Types::PG::JSON do
+  describe 'ROM::SQL::Types::PG::JSON' do
     it 'coerces to pg json hash' do
       input = { foo: 'bar' }
 
@@ -28,10 +27,10 @@ RSpec.describe ROM::SQL::Types, :postgres do
     end
   end
 
-  describe ROM::SQL::Types::PG::Bytea do
+  describe 'ROM::SQL::Types::PG::Bytea' do
     it 'coerses strings to Sequel::SQL::Blob' do
       input = 'sutin'
-      output = described_class[input]
+      output = ROM::SQL::Types::PG::Bytea[input]
 
       expect(output).to be_instance_of(Sequel::SQL::Blob)
       expect(output).to eql('sutin')
